Now implements roles (caregiver or triage_staff) in front-end
