+++ client/src/App.js
@@ -1,17 +1,14 @@
 import React, { Component } from 'react';
-// import axios from 'axios';
 import './App.css';
-// import Cookies from 'universal-cookie';
-// import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import 'antd/dist/antd.css';
 import { Button } from 'antd';
 import Login from './Login.jsx';
 import Admin from './Admin.jsx';
+import Event from './Event.jsx';
 // import $ from 'jquery';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import { Redirect } from "react-router-dom";
 
-// const cookies = new Cookies();
 const LOGGED_IN = 'LOGGED_IN';
 const NOT_LOGGED_IN = 'NOT_LOGGED_IN';
 let jwt_token = localStorage.getItem("token");
@@ -27,11 +24,8 @@ class App extends Component {
     };
   }
 
-
   setUser = email => {
-    this.setState({ email: email }, () =>
-      console.log('Current state after setting user', this.state)
-    );
+    this.setState({ email: email });
     this.setState({ loggedInStatus: LOGGED_IN });
   };
 
@@ -66,9 +60,13 @@ class App extends Component {
               render={(props) => (this.state.loggedInStatus === NOT_LOGGED_IN) ? <Redirect to='/' /> : <Admin {...props} />}
             />
             <Route
+              path="/event"
+              render={(props) => (this.state.loggedInStatus === NOT_LOGGED_IN) ? <Redirect to='/' /> : <Event {...props} />}
+            />
+            <Route
               path="/"
               render={props => (
-                <Login {...props} setUser={this.setUser} setRole={this.setRole} />
+                <Login {...props} setUser={this.setUser} setRole={this.setRole} role={this.state.role} />
               )}
             />
           </Switch>
