+++ client/src/Login.jsx
@@ -21,12 +21,7 @@ class Login extends Component {
     e.preventDefault();
     this.props.form.validateFields((err, values) => {
       if (!err) {
-        // this.props.cookies.set("email", values.email, { path: '/' })
-        // this.props.setUser(this.props.cookies.get('email'))
-        // this.setState({redirect:true})
-        // axios.post('/api/users', {email: this.props.cookies.get('email')})
         const requestJSONobj = { "email": values.email.toLowerCase(), "password": values.password };
-        console.log(requestJSONobj);
         axios.post('api/login', querystring.stringify(requestJSONobj))
           .then((response) => {
             // console.log('Server response', response);
@@ -39,8 +34,8 @@ class Login extends Component {
             if (error.response) {
               // The request was made and the server responded with a status code
               // that falls out of the range of 2xx
-              console.log(error.response.data);
-              console.log(error.response.status);
+              // console.log(error.response.data);
+              // console.log(error.response.status);
               if (error.response.status === 401) {
                 window.alert('Login Error: Incorrect credentials');
               }
@@ -67,7 +62,7 @@ class Login extends Component {
   render() {
     const { getFieldDecorator } = this.props.form;
     if (this.state.redirect) {
-      return (<Redirect to='/admin' />)
+      return this.props.role === 'triage_staff' ? <Redirect to='/admin' /> : <Redirect to='/event' />;
     }
     return (
       <div className="sign-in">
